+++ dynamic-neural-field-composer/include/element_parameters/element_parameters.h
@@ -36,6 +36,9 @@ namespace dnf_composer::element
 		CORRELATED_NORMAL_NOISE_2D,
 		ASYMMETRIC_GAUSS_KERNEL_2D,
 		MEMORY_TRACE_2D,
+
+		RESIZE,
+		RESIZE_2D,
 	};
 
 	inline const std::map<ElementLabel, std::string> ElementLabelToString = {
@@ -65,6 +68,9 @@ namespace dnf_composer::element
 		{CORRELATED_NORMAL_NOISE_2D, "correlated normal noise 2d" },
 		{ASYMMETRIC_GAUSS_KERNEL_2D, "asymmetric gauss kernel 2d" },
 		{MEMORY_TRACE_2D, "memory trace 2d" },
+
+		{RESIZE, "resize" },
+		{RESIZE_2D, "resize 2d" },
 	};
 
 	struct ElementDimensions