+++ dynamic-neural-field-composer/src/elements/element_factory.cpp
@@ -157,6 +157,18 @@ namespace dnf_composer
 					return std::make_shared<MemoryTrace2D>(elementCommonParameters, *params);
 				};
 
+			elementCreators[ElementLabel::RESIZE] = [](const ElementCommonParameters& elementCommonParameters, const ElementSpecificParameters& elementSpecificParameters)
+				{
+					const auto params = dynamic_cast<const ResizeParameters*>(&elementSpecificParameters);
+					return std::make_shared<Resize>(elementCommonParameters, *params);
+				};
+
+			elementCreators[ElementLabel::RESIZE_2D] = [](const ElementCommonParameters& elementCommonParameters, const ElementSpecificParameters& elementSpecificParameters)
+				{
+					const auto params = dynamic_cast<const Resize2DParameters*>(&elementSpecificParameters);
+					return std::make_shared<Resize2D>(elementCommonParameters, *params);
+				};
+
 			}
 
 		std::shared_ptr<Element> ElementFactory::createElement(ElementLabel type, const ElementCommonParameters& elementCommonParameters, const ElementSpecificParameters& elementSpecificParameters)
@@ -226,6 +238,10 @@ namespace dnf_composer
 						return creator->second(ElementCommonParameters(type), AsymmetricGaussKernel2DParameters());
 					case ElementLabel::MEMORY_TRACE_2D:
 						return creator->second(ElementCommonParameters(type), MemoryTrace2DParameters());
+					case ElementLabel::RESIZE:
+						return creator->second(ElementCommonParameters(type), ResizeParameters());
+					case ElementLabel::RESIZE_2D:
+						return creator->second(ElementCommonParameters(type), Resize2DParameters());
 					case ElementLabel::UNINITIALIZED:
 						return nullptr;
 				}